19-11-2025/fraud_detection/monitoring/requirements.txt
kafka-python>=2.0.2
psycopg2-binary>=2.9.9
imbalanced-learn>=0.11.0

# Optional
# python-json-logger==2.0.7  # For structured logging
# prometheus-client==0.19.0  # For metrics export